Remove "mock_" prefix from mocks in fakes/
- eliminates redundancy that makes names long
- also remove unnecessary struct property initialization to default
  values and remove line breaks in imports
