+++ pkg/fakes/crypto_parser.go
