+++ pkg/fakes/header_sync_header_repository.go
@@ -1,9 +1,24 @@
+// VulcanizeDB
+// Copyright © 2019 Vulcanize
+
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU Affero General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU Affero General Public License for more details.
+
+// You should have received a copy of the GNU Affero General Public License
+// along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
 package fakes
 
 import "github.com/vulcanize/vulcanizedb/pkg/core"
 
-type MockHeaderSyncHeaderRepository struct {
-}
+type MockHeaderSyncHeaderRepository struct{}
 
 func (*MockHeaderSyncHeaderRepository) AddCheckColumn(id string) error {
 	return nil
