+++ pkg/fakes/header_sync_log_repository.go
