--- pkg/fakes/mock_full_sync_block_retriever.go
@@ -1,15 +1,0 @@
-package fakes
-
-type MockFullSyncBlockRetriever struct {
-	FirstBlock      int64
-	FirstBlockErr   error
-	MostRecentBlock int64
-}
-
-func (retriever *MockFullSyncBlockRetriever) RetrieveFirstBlock(contractAddr string) (int64, error) {
-	return retriever.FirstBlock, retriever.FirstBlockErr
-}
-
-func (retriever *MockFullSyncBlockRetriever) RetrieveMostRecentBlock() (int64, error) {
-	return retriever.MostRecentBlock, nil
-}
